presets/api/__global__/requirements.txt
Django==4.2.7
django-cleanup
django-cors-headers
django-debug-toolbar
django-filter
django-import-export
django-model-utils
django-simple-history
django-tinymce
django-unfold
djangorestframework
djangorestframework-simplejwt
gunicorn
numpy
packaging
pandas
pillow
PyJWT
python-dateutil
pytz
six
sqlparse
tablib
typing_extensions
tzdata
